backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
celery>=5.3.0
redis>=5.0.0
beautifulsoup4>=4.12.0
scrapy>=2.11.0
playwright>=1.40.0
newspaper3k>=0.2.8
pypandoc>=1.12
weasyprint>=60.0
ebooklib>=0.18
pypdf2>=3.0.0
python-docx>=1.1.0
requests>=2.31.0
pillow>=10.0.0
langdetect>=1.0.9
python-multipart>=0.0.6
aiofiles>=23.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
lxml>=4.9.0
html5lib>=1.1
markdown>=3.5.0
pyyaml>=6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
